pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "asr_benchmark"
version = "0.0.1"
description = "Benchmarks for the On Top of Pasketti: Children\u2019s Speech Recognition Challenge."
authors = [
  { name = "DrivenData" },
]
license = { file = "LICENSE" }
readme = "README.md"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License"
]
requires-python = "~=3.11.13"
dependencies = [
    "cuda-python>=12.6,<12.7",
    "datasets",
    "ipykernel",
    "ipywidgets",
    "jiwer",
    "librosa",
    "loguru",
    "matplotlib",
    "nbautoexport",
    "nbconvert",
    "nemo_toolkit[all]",
    "papermill",
    "pandas",
    "pip",
    "soundfile",
    "transformers[torch]>=4.52.4",
    "torch>=2.9.0",
    "torchaudio",
    "torchcodec",
    "tqdm",
    "ruff",
]


[tool.ruff]
line-length = 99
src = ["asr_benchmark", "phonetic_submission"]
include = ["pyproject.toml", "asr_benchmark/**/*.py", "phonetic_submission/**/*.py"]

[tool.ruff.lint]
extend-select = ["I"]  # Add import sorting

[tool.ruff.lint.isort]
known-first-party = ["asr_benchmark"]
force-sort-within-sections = true

[tool.uv.sources]
torch = { index = "pytorch-cu126" }
torchaudio = { index = "pytorch-cu126" }

[[tool.uv.index]]
name = "pytorch-cu126"
url = "https://download.pytorch.org/whl/cu126"
explicit = true
